Update NAS2D submodule to most recent version
Why not. There is nothing that breaks, which isn't already broken.

Currently there is no `data/` folder, and the tests projects reference
non-existent resources from that folder, so that is broken. I guess
technically it was always broken, though previously that was just a
warning, and empty placeholders were used. During updates to the Color
rename, NAS2D changes were brought in that increased the warnings to
errors.