Remove Zombie constructor overload for integers
The constructor parameter types should match the internally used types.
Any casting to convert should be done outside the class.

Previously, the integer constructor was being called with unsigned
integers, so implicit conversion was still being done outside the class
anyway. This lead to warnings on high warning levels.

The deeper problem here seems to be a lack of consensus on whether
internal values should be float or int. Both are used in a somewhat
inconsistent manner.