Avoid undefined behaviour by using std::erase + std::remove_if
The previous code could result in undefined behavior, since `erase`
would invalidate iterators at or beyond the element being erased. At
best, this would skip over the next entry once it was moved back to the
current position, and at worst, if updating the code to use iterators,
it may skip over the `.end()` element and cause a crash.

One of the reasons the old code was able to function, was the
impossibility of killing two zombies at the same time, and so the code
never needed to remove two consecutive zombies during the same loop.

The other reason the old code didn't break, was the inequality, and that
vectors didn't generally shrink in size, so you would never access
memory outside of the vector, even if the end point has moved due to
elements being erased. With iterators, a strict not equal check is used,
which might not terminate the loop if you skip over the end element.